+++ src/app/shared/models/characters/Character.ts
@@ -1,3 +1,7 @@
+import { Url } from "../common/common"; 
+import { SeriesSummary } from "../common/common";
+import { Image } from "../common/common";
+
 export interface Character{
     id:number;
     name:string;
@@ -29,16 +33,6 @@ export interface CharacterDataWrapper{
     etag:string;
 }
 
-export interface Url{
-    type:string;
-    url:string;
-}
-
-export interface SeriesSummary{
-    resourceURI:string;
-    name:string;
-}
-
 export interface SeriesList{
     available:number;
     returned:number;
@@ -46,10 +40,7 @@ export interface SeriesList{
     items:SeriesSummary[];
 }
 
-export interface Image{
-    path:string;
-    extension:string;
-}
+
 
 export interface ComicList{
     available:number;
