+++ src/app/shared/models/common/common.ts
@@ -1,0 +1,14 @@
+export interface Url{
+    type : string;
+    url : string;
+}
+
+export interface SeriesSummary{
+    resourceURI:string;
+    name:string;
+}
+
+export interface Image{
+    path : string
+    extension : string
+}
